requirements.txt
conllu==4.3
joblib==1.0.0
numpy==1.19.5
PyYAML==5.4.1
scikit-learn==0.24.1
scipy==1.6.0
sklearn==0.0
threadpoolctl==2.1.0
torch==1.7.1
typing-extensions==3.7.4.3
